+++ Userland/SampleCodeModule/processes/processes.c
@@ -1,38 +1,38 @@
 #include <stdio.h>
 
 void processKiller(){
-    uint64_t pid;
+    int pid;
     getPIDSyscall(&pid);
     int ans;
     killSyscall(pid,&ans);
     while (1);
 }
 
-uint64_t getPid(){
-    uint64_t pid;
+int getPid(){
+    int pid;
     getPIDSyscall(&pid);
     return pid;
 }
 
-int nice(uint64_t pid, uint64_t prio){
+int nice(int pid, int prio){
     int ans;
     niceSyscall(pid, prio, &ans);
     return ans;
 }
 
-int block(uint64_t pid) {
+int block(int pid) {
     int ans;
     blockSyscall(pid, &ans); //falta syscall
     return ans;
 }
 
-int unblock(uint64_t pid) {
+int unblock(int pid) {
     int ans;
     unblockSyscall(pid, &ans); // falta syscall
     return ans;
 }
 
-int kill(uint64_t pid){
+int kill(int pid){
     int ans;
     killSyscall(pid, &ans);
     return ans;
